test: 테스트 중복 제거 및 올바른 객체 테스트로 이동
