+++ src/test/java/model/BoardTest.java
@@ -1,15 +1,15 @@
 package model;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import model.coordinate.Position;
 import model.piece.Chariot;
 import model.piece.Piece;
 import org.junit.jupiter.api.Test;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class BoardTest {
 
@@ -55,5 +55,59 @@ class BoardTest {
 
         // then
         assertThat(board.pickPiece(next)).isEqualTo(han);
+    }
+
+    @Test
+    void 이동_경로에_기물이_있으면_멱이_적용된다() {
+        // given
+        Piece chariot = new Chariot(Team.HAN);
+        Position current = new Position(0, 0);
+        Position next = new Position(0, 5);
+
+        Map<Position, Piece> pieces = new HashMap<>();
+        pieces.put(current, chariot);
+        pieces.put(new Position(0, 3), new Chariot(Team.CHO));
+        Board board = new Board(pieces);
+
+        // when
+        List<Position> path = chariot.extractPath(current, next);
+
+        // then
+        assertThat(board.hasPieceAt(path)).isTrue();
+    }
+
+
+    @Test
+    void 한_칸_이동은_경로가_없으므로_멱이_적용되지_않는다() {
+        // given
+        Piece chariot = new Chariot(Team.HAN);
+        Position current = new Position(0, 0);
+        Position next = new Position(0, 1);
+
+        Board board = new Board(new HashMap<>());
+
+        // when
+        List<Position> path = chariot.extractPath(current, next);
+
+        // then
+        assertThat(board.hasPieceAt(path)).isFalse();
+    }
+
+    @Test
+    void 이동_경로에_기물이_없으면_멱이_적용되지_않는다() {
+        // given
+        Piece chariot = new Chariot(Team.HAN);
+        Position current = new Position(0, 0);
+        Position next = new Position(0, 5);
+
+        Map<Position, Piece> pieces = new HashMap<>();
+        pieces.put(current, chariot);
+        Board board = new Board(pieces);
+
+        // when
+        List<Position> path = chariot.extractPath(current, next);
+
+        // then
+        assertThat(board.hasPieceAt(path)).isFalse();
     }
 }
