+++ src/test/java/model/piece/ChariotTest.java
@@ -2,10 +2,7 @@ package model.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import model.Board;
 import model.Team;
 import model.coordinate.Position;
 import org.junit.jupiter.api.Test;
@@ -27,7 +24,8 @@ public class ChariotTest {
 
     @ParameterizedTest
     @MethodSource("model.fixture.PieceTestFixture#사간방_대각선_이동_방향_케이스")
-    void 차는_대각선으로_이동할_수_없다(Position current, Position next) {
+    @MethodSource("model.fixture.PieceTestFixture#제자리_이동_케이스")
+    void 차는_대각선_또는_제자리로_이동할_수_없다(Position current, Position next) {
         // given
         Piece chariot = new Chariot(Team.HAN);
         // when
@@ -68,58 +66,4 @@ public class ChariotTest {
         // then
         assertThat(path).isEmpty();
     }
-
-    @Test
-    void 이동_경로에_기물이_있으면_멱이_적용된다() {
-        // given
-        Piece chariot = new Chariot(Team.HAN);
-        Position current = new Position(0, 0);
-        Position next = new Position(0, 5);
-
-        Map<Position, Piece> pieces = new HashMap<>();
-        pieces.put(current, chariot);
-        pieces.put(new Position(0, 3), new Chariot(Team.CHO));
-        Board board = new Board(pieces);
-
-        // when
-        List<Position> path = chariot.extractPath(current, next);
-
-        // then
-        assertThat(board.hasPieceAt(path)).isTrue();
-    }
-
-    @Test
-    void 이동_경로에_기물이_없으면_멱이_적용되지_않는다() {
-        // given
-        Piece chariot = new Chariot(Team.HAN);
-        Position current = new Position(0, 0);
-        Position next = new Position(0, 5);
-
-        Map<Position, Piece> pieces = new HashMap<>();
-        pieces.put(current, chariot);
-        Board board = new Board(pieces);
-
-        // when
-        List<Position> path = chariot.extractPath(current, next);
-
-        // then
-        assertThat(board.hasPieceAt(path)).isFalse();
-    }
-
-    @Test
-    void 한_칸_이동은_경로가_없으므로_멱이_적용되지_않는다() {
-        // given
-        Piece chariot = new Chariot(Team.HAN);
-        Position current = new Position(0, 0);
-        Position next = new Position(0, 1);
-
-        Board board = new Board(new HashMap<>());
-
-        // when
-        List<Position> path = chariot.extractPath(current, next);
-
-        // then
-        assertThat(board.hasPieceAt(path)).isFalse();
-    }
-
 }
