+++ src/test/java/model/piece/SoldierTest.java
@@ -1,11 +1,11 @@
 package model.piece;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import model.Team;
 import model.coordinate.Position;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 public class SoldierTest {
 
@@ -14,8 +14,10 @@ public class SoldierTest {
     void 한나라_병은_전진과_좌우로_이동할_수_있다(Position current, Position next) {
         // given
         Piece soldier = new Soldier(Team.HAN);
+
         // when
         boolean canMove = soldier.canMove(current, next);
+
         // then
         assertThat(canMove).isTrue();
     }
@@ -25,8 +27,10 @@ public class SoldierTest {
     void 한나라_병은_후퇴와_두칸_이동을_할_수_없다(Position current, Position next) {
         // given
         Piece soldier = new Soldier(Team.HAN);
+
         // when
         boolean canMove = soldier.canMove(current, next);
+
         // then
         assertThat(canMove).isFalse();
     }
@@ -36,8 +40,10 @@ public class SoldierTest {
     void 초나라_졸은_전진과_좌우로_이동할_수_있다(Position current, Position next) {
         // given
         Piece soldier = new Soldier(Team.CHO);
+
         // when
         boolean canMove = soldier.canMove(current, next);
+
         // then
         assertThat(canMove).isTrue();
     }
@@ -47,8 +53,10 @@ public class SoldierTest {
     void 초나라_졸은_후퇴와_두칸_이동을_할_수_없다(Position current, Position next) {
         // given
         Piece soldier = new Soldier(Team.CHO);
+
         // when
         boolean canMove = soldier.canMove(current, next);
+
         // then
         assertThat(canMove).isFalse();
     }
